requirements.txt
feedparser>=6.0.10
httpx>=0.27.0
openai>=1.12.0
beautifulsoup4>=4.12.3
Pillow>=10.2.0
python-dateutil>=2.8.2
pydantic>=2.6.0
tenacity>=8.2.3
